refactor(propType): 💡 rename propType
