+++ src/lib/propType.js
